delete category

+++ php/public/admin/pages/categoryList.php
@@ -55,7 +55,7 @@ $categories = $categoryCont->getAllCategory();
                                         <td><img src="<?= "images/" . $category["image"] ?>" style="width:50px; aspect-ratio:1/1; object-fit:cover;" alt=""></td>
                                         <td><?= $category["title"] ?></td>
                                         <td>Edit</td>
-                                        <td><a href="/../includes/category-delete.php?id=<?= $category["id"] ?>">Delete</a></td>
+                                        <td><a href="/../includes/category-delete.php?id=<?= $category["id"] ?>" onclick="return confirm('Are you sure to delete category?')">Delete</a></td>
                                     </tr>
                                 <?php endforeach; ?>
